+++ changelog.md
@@ -8,6 +8,14 @@ Team membership:
 Team conventions: Allman notation, markdown for changelog  
 Changelog format: [Markdown](https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet) 
 
+ ## [0.0.4] - 2018-02-09
+ ### Added
+ - Equipment and preset render
+ - Equipment/preset controller and data array
+
+ ### Updated
+ - changelog
+
 ## [0.0.3] - 2018-02-09
 ### Added
 - Added placeholder pages for assembly, history, and about